backend/pyproject.toml
[project]
name = "ranktastic-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.1.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "apscheduler>=3.10",
    "bcrypt>=4.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "greenlet>=3.3.2",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
